Revised the fix for ticket #847, to be similar to the fix which is in the /python3 branch [2272].
This fix is slightly different from the one in /python3.  It only sends a 408 if the timeout occurs while in the middle of a request.  When no request has been seen on the connection, it does not send a 408; it just closes the connection.

--HG--
extra : convert_revision : svn%3Ae1d34091-3ce9-0310-8e96-997e60db3bd5/trunk%402281