Convert the tests to use nose instead of our own runner.  This strips out much coverage and profiling (handled by nose) and lets you focus on writing tests.
The biggest changes that have to be done in the tests classes is you have to put the "setup_server" method on the class(es) that need them when running.  If you need it for multiple classes, you can use staticmethod() to attach it to multiple classes without using inheritance.

--HG--
extra : convert_revision : svn%3Ae1d34091-3ce9-0310-8e96-997e60db3bd5/trunk%402637