Make SCRIPT_NAME environ variable optional
PEP 333 makes this CGI-style variable optional.  See the the section titled "environ variables":

> The following variables must be present, unless their value would be an empty string, in which case they may be omitted, except as otherwise noted below.

Some WSGI middleware does not set this environ variable.  When combined with `cheroot.wsgi.PathInfoDispatcher`, there is a confusing `KeyError` exception that may not be correctly be translated to an HTTP 500 response.

This change makes the variable optional.