Be more graceful when terminating keep-alive connections. Ref #263.
The previous behaviour was that when we exceed our threshold of permitted
keep-alive connections, we would evict the least recently used connection
by forcibly shutting down the socket. This would cause problems with clients
which wouldn't be expecting socket errors.

Now, if we have a connection that we would usually keep-alive, but we have
already reached our limit of allowed keep-alive connections, then we close
the connection gracefully by sending a "Connection: Close" header for
HTTP/1.1 or omitting the "Connection" header for HTTP/1.0.

This has the downside of having cheroot holding on to connections which are
less recently used, rather than the most recent ones - but it does ensure that
we make the decision about whether to keep or drop a connection at the time we
are writing the headers, which allows the client to have a better expectation
about the state of the socket once the response has been read.

However, we still forcibly close sockets for idle keep-alive connections once
the server timeout has been exceeded.