Implementa a DefaultSelector manager for the ConnectionManager.
The new wrappers keeps an additional copy of the selected
sockets, the external copy allow is to concurrently
iterate and modify the list of the selected sockets.

In addition, fixes an issue when the server is never notified
that is not serving the the process gets interrupted inside
a `safe_start` call and `stop` gets stuck in a while loop,
waiting to be ready to exit gracefully.

Fixes issue #313.