Test improvements related to issue 317.
Include the server error_log verification in the tests defined in
test_wsgi, this test in particular seems to be good enough
to start to trigger the race condition. Is not deterministic,
but at least its start.

In addition remove the exclusion of the special case for python2
in test_conn.test_invalid_selected_connections, the error it was
covering other issues with the differences on how the
HTTPConnection is closed in Python2 vs Python3. It had a comment
about how _old_ versions of Python didn't close the OS socket,
this is in a way, still true, even for python3, it will only close
the socket until there are no other references to it. Instead of
issuing the close to the kernel socket, it issues a 'shutdown',
to guarantee that the tcp connection is closed, and then it
tries to close the socket to free the file descriptor.