🚑 Fix test_high_number_of_file_descriptors @ PyPy
This test seems to be flakily failing under PyPy in CIs. The working
theory is that this is happening because of PyPy's indeterministic GC
mechanism that releases file descriptors right before your test run
which causes lower file descriptor numbers to be reallocated to the
server process which is not fit for reproducing the regression.