connections: expiration_interval as select timeout
This is a continuation of: https://github.com/cherrypy/cheroot/pull/352

As discussed in that pull request, on Windows the select() method does not return when a socket is ready. While the reason is still to be found out, to get the benefit of using a usually much higher timeout merged, Windows is not handled differently: The timeout is capped to 0.05 seconds to assure that connections are not delayed more than that.

Signed-off-by: MichaIng <micha@dietpi.com>