Drop broken Dependabot integration for now
At the moment, Dependabot is only able to use one Python interpreter
version per-project which is not what we need. The inability to
support our use-case results in crashes with tracebacks appearing in
the Dependabot log and it just does nothing.

This is why, we're disabling this integration until the bug is fixed
upstream.

Refs:
* https://github.com/dependabot/dependabot-core/issues/4606
* https://github.com/dependabot/dependabot-core/issues/4607