Use importlib.metadata instead of deprecated pkg_resources
pkg_resources is deprecated upstream and does not handle modern
.dist-info directories correctly. Specifically it does not honor the new
replacement rules as defined in [1]. As a result, it can't find some of
cheroot's dependencies on modern Gentoo systems and ultimately fails to
load with a DistributionNotFound error [2].

pkg_resources is only used to get the cheroot version and can easily be
replaced with importlib.metadata [3], the intended replacement for this
specific use-case. For python versions before 3.8, attempt to use the
importlib_metadata backport instead.

[1] https://packaging.python.org/en/latest/specifications/binary-distribution-format/#escaping-and-unicode
[2] https://bugs.gentoo.org/834522
[3] https://docs.python.org/3/library/importlib.metadata.html

Thanks-to: Michał Górny <mgorny@gentoo.org>
Signed-off-by: Wolfgang Müller <wolf@oriole.systems>