Handle SSLZeroReturnError exceptions
Python 3.8 introduced a different exception for zero byte tcp
connections.  These connections are generated by cherrypy on
startup, and so the exception is not displayed as it is expected.
Without this fix an exception is displayed although it is harmless.
This treats the new exception just like it was treated under the
errno exception.