Switch the commit message check to `gitlint` in CI
This patch replaces the previous broken implementation [1] that was
using the `platisd/bad-commit-message-blocker` GitHub action with
manual calls of the `gitlint` tool.
The new implementation also attempts to expose the details of what is
being linted and adds the results to the workflow job summary. It
retrieves the number of the PR commits but no more than 50 and makes
sure to fetch all of them. If there's more, the workflow fails the
check as well.

[[1]]: https://github.com/platisd/bad-commit-message-blocker/issues/12