🐛 Turn `SSLEOFError` into `FatalSSLAlert`
This patch turns a new `ssl.SSLEOFError` into an internally ignored
`FatalSSLAlert` allowing it not to leak into the outer abstraction
layers in its raw form.
The exception is new since Python 3.8 and it's fine to use it
unconditionally since we no longer support Python 3.7.

This patch also handles `SSLZeroReturnError` same as `SSLEOFError`
as it's semantically equivalent per [[1]].

[1]: https://github.com/cherrypy/cheroot/pull/518#issuecomment-1631774708