♲ Stop handling `EINTR` around `time.sleep()`
`IOError` is an alias of `OSError` since Python 3.3. Python 3 also
indroduced a new exception `InterruptedError` which represents
`EINTR`. The `time.sleep()` call could raise

    IOError: [Errno 4] Interrupted function call

on KBInt under Python 2, which would be `InterruptedError` under
Python 3 but it's not raised anymore post PEP 475 that was implemented
in Python 3.5. So it does not actually need to be handled in modern
runtimes.

Refs:
* https://stackoverflow.com/a/52613818/595220
* https://peps.python.org/pep-0475/
* https://github.com/python/cpython/issues/56671
* https://stackoverflow.com/a/38258781/595220
* https://docs.python.org/3/library/exceptions.html#InterruptedError
* https://github.com/python/cpython/issues/36893