🐛🧪 Fix the metadata verification cmd @ tox
Due to a missing EOL escaping, tox was treating the following line
as a separate command. But since it was starting with a CLI option
having a dash in front of it, tox decided that the return code of this
command is requested to be ignored. This was a hard to catch case.