Daemonize the serve_unservicable thread
In Python < 3.13, the code backfills a shutdown method for a queue, but
this isn't complete; an existing get call is not interrupted by shutdown
and this means that when the unservicable queue is empty, it never
stops, preventing the system from stopping.  By daemonizing the
unservicable thread, the server can stop and exit.

A better solution would probably involve copying a fair amount of code
from Python 3.13's queue implementation.  In my opinion that isn't worth
the maintenance burden.  If desired, perhaps we should only daemonize
this thread on Python < 3.13.

Signed-off-by: David Manthey <david.manthey@kitware.com>