Align SSL adapters by removing `bind()`
Relocated `SSLConnection` instantiation from `bind()` to `wrap()` to align
with the adapter interface where `wrap()` is responsible for converting
raw sockets to SSL connections. With this chenge `bind()` serves no
purpose and so is deprecated. This change also fixes the pyOpenSSL adapter
so that it doesn't prematurely wrap the raw socket, an issue in the old
`bind()` implementation.